Quick fix (28/02/22)
The positioning of the labels for the tables caused them to appear as plain text, so they have been shifted down. Similarly, the examples given for fractional numerals were not indented, so they were not appearing as code blocks. Lastly, one of the labels that was not italicized has been italicized